+++ Development-Team/example/Agent-AntiCheat/Agent_AntiCheat.vcxproj
@@ -23,7 +23,7 @@
     <Keyword>Win32Proj</Keyword>
     <ProjectGuid>{1abb64a7-3e6e-43b0-a11d-01748f1a2bc9}</ProjectGuid>
     <RootNamespace>AgentAntiCheat</RootNamespace>
-    <WindowsTargetPlatformVersion>10.0</WindowsTargetPlatformVersion>
+    <WindowsTargetPlatformVersion>10.0.19041.0</WindowsTargetPlatformVersion>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'" Label="Configuration">
@@ -42,7 +42,7 @@
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'" Label="Configuration">
     <ConfigurationType>Application</ConfigurationType>
     <UseDebugLibraries>true</UseDebugLibraries>
-    <PlatformToolset>v142</PlatformToolset>
+    <PlatformToolset>v143</PlatformToolset>
     <CharacterSet>Unicode</CharacterSet>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'" Label="Configuration">
@@ -78,6 +78,7 @@
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <LinkIncremental>true</LinkIncremental>
+    <ExecutablePath>C:\Program Files (x86)\Windows Kits\10\bin\10.0.19041.0\x64;$(ExecutablePath)</ExecutablePath>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <LinkIncremental>false</LinkIncremental>
@@ -141,21 +142,28 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClInclude Include="Debugutil.h" />
     <ClInclude Include="main.h" />
     <ClInclude Include="framework.h" />
     <ClInclude Include="Resource.h" />
     <ClInclude Include="targetver.h" />
     <ClInclude Include="driverport.h" />
+    <ClInclude Include="Tray.h" />
+    <ClInclude Include="Winproc.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="Debugutil.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="DriverPort.cpp" />
+    <ClCompile Include="Tray.cpp" />
+    <ClCompile Include="Winproc.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="Agent_AntiCheat.rc" />
   </ItemGroup>
   <ItemGroup>
     <Image Include="Agent_AntiCheat.ico" />
+    <Image Include="bitmap1.bmp" />
     <Image Include="small.ico" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
