+++ Development-Team/example/Agent-AntiCheat/Agent_AntiCheat.vcxproj.filters
@@ -13,30 +13,69 @@
       <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
       <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
     </Filter>
+    <Filter Include="소스 파일\DriverPort">
+      <UniqueIdentifier>{d353e914-ba90-4e05-9567-6cbc27dd61fd}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="소스 파일\Main">
+      <UniqueIdentifier>{285c69ef-cd30-4523-a20d-ef08fed42ea3}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="소스 파일\UI">
+      <UniqueIdentifier>{9bfce389-8027-42cc-921e-04ce1463e3dd}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="헤더 파일\DriverPort">
+      <UniqueIdentifier>{c3e0ef54-324b-4c7c-af69-8d0650d43bea}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="헤더 파일\Main">
+      <UniqueIdentifier>{59b28b21-2edb-4d34-8102-325d1a4c7090}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="헤더 파일\Common">
+      <UniqueIdentifier>{0b169fd8-bf78-479d-ba2a-b2d19dece532}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="헤더 파일\UI">
+      <UniqueIdentifier>{345c8a03-fb85-45b7-ae6a-06f71ccee55a}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="framework.h">
-      <Filter>헤더 파일</Filter>
+    <ClInclude Include="driverport.h">
+      <Filter>헤더 파일\DriverPort</Filter>
     </ClInclude>
-    <ClInclude Include="targetver.h">
-      <Filter>헤더 파일</Filter>
+    <ClInclude Include="framework.h">
+      <Filter>헤더 파일\Common</Filter>
     </ClInclude>
     <ClInclude Include="Resource.h">
-      <Filter>헤더 파일</Filter>
+      <Filter>헤더 파일\Common</Filter>
+    </ClInclude>
+    <ClInclude Include="targetver.h">
+      <Filter>헤더 파일\Common</Filter>
     </ClInclude>
     <ClInclude Include="main.h">
-      <Filter>헤더 파일</Filter>
+      <Filter>헤더 파일\Main</Filter>
     </ClInclude>
-    <ClInclude Include="driverport.h">
-      <Filter>헤더 파일</Filter>
+    <ClInclude Include="Tray.h">
+      <Filter>헤더 파일\UI</Filter>
+    </ClInclude>
+    <ClInclude Include="Winproc.h">
+      <Filter>헤더 파일\UI</Filter>
+    </ClInclude>
+    <ClInclude Include="Debugutil.h">
+      <Filter>헤더 파일\DriverPort</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="main.cpp">
-      <Filter>소스 파일</Filter>
-    </ClCompile>
     <ClCompile Include="DriverPort.cpp">
-      <Filter>소스 파일</Filter>
+      <Filter>소스 파일\DriverPort</Filter>
+    </ClCompile>
+    <ClCompile Include="main.cpp">
+      <Filter>소스 파일\Main</Filter>
+    </ClCompile>
+    <ClCompile Include="Tray.cpp">
+      <Filter>소스 파일\UI</Filter>
+    </ClCompile>
+    <ClCompile Include="Winproc.cpp">
+      <Filter>소스 파일\UI</Filter>
+    </ClCompile>
+    <ClCompile Include="Debugutil.cpp">
+      <Filter>소스 파일\DriverPort</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -49,6 +88,9 @@
       <Filter>리소스 파일</Filter>
     </Image>
     <Image Include="Agent_AntiCheat.ico">
+      <Filter>리소스 파일</Filter>
+    </Image>
+    <Image Include="bitmap1.bmp">
       <Filter>리소스 파일</Filter>
     </Image>
   </ItemGroup>
