+++ Development-Team/example/Agent-AntiCheat/Debugutil.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <string>
+
+void DebugPrintA(const char* format, ...);
+
+void DebugPrintW(const wchar_t* format, ...);
