+++ Development-Team/example/Agent-AntiCheat/Resource.h
@@ -1,30 +1,33 @@
 ﻿//{{NO_DEPENDENCIES}}
 // Microsoft Visual C++에서 생성한 포함 파일입니다.
-// 다음에서 사용 Agent_AntiCheat.rc
-
-#define IDS_APP_TITLE			103
-
-#define IDR_MAINFRAME			128
-#define IDD_AGENTANTICHEAT_DIALOG	102
-#define IDD_ABOUTBOX			103
-#define IDM_ABOUT				104
-#define IDM_EXIT				105
-#define IDI_AGENTANTICHEAT			107
-#define IDI_SMALL				108
-#define IDC_AGENTANTICHEAT			109
-#define IDC_MYICON				2
-#ifndef IDC_STATIC
-#define IDC_STATIC				-1
-#endif
-// 다음은 새 개체에 사용할 기본값입니다.
+// Agent_AntiCheat.rc에서 사용되고 있습니다.
 //
+#define IDC_MYICON                      2
+#define IDD_AGENTANTICHEAT_DIALOG       102
+#define IDS_APP_TITLE                   103
+#define IDD_ABOUTBOX                    103
+#define IDM_ABOUT                       104
+#define IDM_EXIT                        105
+#define IDI_AGENTANTICHEAT              107
+#define IDI_SMALL                       108
+#define IDC_AGENTANTICHEAT              109
+#define IDR_MAINFRAME                   128
+#define IDB_BITMAP1                     130
+#define IDC_EDIT_PATH                   1010
+#define IDC_BUTTON4                     1011
+#define IDC_BUTTON_BROWSE               1011
+#define IDOK                            1012
+#define IDC_BUTTON1                     1013
+#define IDC_STATIC                      -1
+
+// Next default values for new objects
+// 
 #ifdef APSTUDIO_INVOKED
 #ifndef APSTUDIO_READONLY_SYMBOLS
-
-#define _APS_NO_MFC					130
-#define _APS_NEXT_RESOURCE_VALUE	129
-#define _APS_NEXT_COMMAND_VALUE		32771
-#define _APS_NEXT_CONTROL_VALUE		1000
-#define _APS_NEXT_SYMED_VALUE		110
+#define _APS_NO_MFC                     1
+#define _APS_NEXT_RESOURCE_VALUE        132
+#define _APS_NEXT_COMMAND_VALUE         32771
+#define _APS_NEXT_CONTROL_VALUE         1014
+#define _APS_NEXT_SYMED_VALUE           110
 #endif
 #endif
