+++ Development-Team/example/Agent-AntiCheat/Winproc.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "framework.h"
+
+LRESULT CALLBACK WndProc(HWND, UINT, WPARAM, LPARAM);
+INT_PTR CALLBACK About(HWND, UINT, WPARAM, LPARAM);
