+++ Development-Team/example/Agent-AntiCheat/framework.h
@@ -5,19 +5,20 @@
 #pragma once
 
 #include "targetver.h"
+#include "resource.h"
+
 #define WIN32_LEAN_AND_MEAN             // 거의 사용되지 않는 내용을 Windows 헤더에서 제외합니다.
 // Windows 헤더 파일
 #include <windows.h>
-#include <Windows.h>
 // C 런타임 헤더 파일입니다.
 #include <stdlib.h>
 #include <malloc.h>
 #include <memory.h>
 #include <tchar.h>
 
-#include <shellapi.h>
 
+#include <shellapi.h>
+#include <commdlg.h>
 
 //DriverPort
 #include <fltUser.h>
-#include "driverport.h"
