+++ Development-Team/example/Agent-AntiCheat/main.cpp
@@ -1,17 +1,11 @@
 ﻿#include "framework.h"
+
 #include "main.h"
 
-#define MAX_LOADSTRING 100
-#define WM_TRAYICON (WM_USER + 1)
-
-HINSTANCE hInst;
 NOTIFYICONDATA nid = { 0 };
+HINSTANCE hInst;
 HWND hWnd;
 
-LRESULT CALLBACK WndProc(HWND, UINT, WPARAM, LPARAM);
-INT_PTR CALLBACK About(HWND, UINT, WPARAM, LPARAM);
-void InitTrayIcon(HWND);
-void ShowContextMenu(HWND, POINT);
 
 int APIENTRY wWinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPWSTR lpCmdLine, int nCmdShow)
 {
@@ -43,77 +37,4 @@ int APIENTRY wWinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPWSTR lpCmd
 
     Shell_NotifyIcon(NIM_DELETE, &nid);
     return (int)msg.wParam;
-}
-
-void InitTrayIcon(HWND hWnd)
-{
-    nid.cbSize = sizeof(NOTIFYICONDATA);
-    nid.hWnd = hWnd;
-    nid.uID = 1;
-    nid.uFlags = NIF_ICON | NIF_MESSAGE | NIF_TIP;
-    nid.uCallbackMessage = WM_TRAYICON;
-    nid.hIcon = LoadIcon(hInst, MAKEINTRESOURCE(IDI_AGENTANTICHEAT));
-    wcscpy_s(nid.szTip, L"Agent_AntiCheat");
-    Shell_NotifyIcon(NIM_ADD, &nid);
-}
-
-void ShowContextMenu(HWND hWnd, POINT pt)
-{
-    HMENU hMenu = CreatePopupMenu();
-    InsertMenu(hMenu, -1, MF_BYPOSITION, IDM_ABOUT, L"정보(&I)");
-    InsertMenu(hMenu, -1, MF_BYPOSITION, IDM_EXIT, L"끝내기(&X)");
-
-    SetForegroundWindow(hWnd); // 메뉴가 제대로 동작하도록
-
-    TrackPopupMenu(hMenu, TPM_RIGHTBUTTON, pt.x, pt.y, 0, hWnd, nullptr);
-    DestroyMenu(hMenu);
-}
-
-LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
-{
-    switch (message)
-    {
-    case WM_TRAYICON:
-        if (lParam == WM_RBUTTONUP || lParam == WM_LBUTTONUP)
-        {
-            POINT pt;
-            GetCursorPos(&pt);
-            ShowContextMenu(hWnd, pt);
-        }
-        break;
-    case WM_COMMAND:
-        switch (LOWORD(wParam))
-        {
-        case IDM_ABOUT:
-            DialogBox(hInst, MAKEINTRESOURCE(IDD_ABOUTBOX), hWnd, About);
-            break;
-        case IDM_EXIT:
-            DestroyWindow(hWnd);
-            break;
-        }
-        break;
-    case WM_DESTROY:
-        PostQuitMessage(0);
-        break;
-    default:
-        return DefWindowProc(hWnd, message, wParam, lParam);
-    }
-    return 0;
-}
-
-INT_PTR CALLBACK About(HWND hDlg, UINT message, WPARAM wParam, LPARAM lParam)
-{
-    UNREFERENCED_PARAMETER(lParam);
-    if (message == WM_INITDIALOG)
-        return (INT_PTR)TRUE;
-
-    if (message == WM_COMMAND)
-    {
-        if (LOWORD(wParam) == IDOK || LOWORD(wParam) == IDCANCEL)
-        {
-            EndDialog(hDlg, LOWORD(wParam));
-            return (INT_PTR)TRUE;
-        }
-    }
-    return (INT_PTR)FALSE;
 }
